adiciona scripts para geracao dicionarios de dados
