+++ requirements.in
@@ -1,1 +1,1 @@
-frictionless[excel]
+frictionless[excel, pandas]
