+++ requirements.txt
@@ -1,8 +1,8 @@
 #
-# This file is autogenerated by pip-compile with Python 3.12
+# This file is autogenerated by pip-compile with Python 3.11
 # by the following command:
 #
-#    pip-compile requirements.in
+#    pip-compile
 #
 annotated-types==0.7.0
     # via pydantic
@@ -19,11 +19,9 @@ charset-normalizer==3.3.2
     # via requests
 click==8.1.7
     # via typer
-colorama==0.4.6
-    # via click
 et-xmlfile==1.1.0
     # via openpyxl
-frictionless[excel]==5.17.0
+frictionless[excel,pandas]==5.17.0
     # via -r requirements.in
 humanize==4.10.0
     # via frictionless
@@ -47,9 +45,17 @@ markupsafe==2.1.5
     # via jinja2
 mdurl==0.1.2
     # via markdown-it-py
+numpy==2.1.0
+    # via
+    #   pandas
+    #   pyarrow
 openpyxl==3.1.5
     # via frictionless
+pandas==2.2.2
+    # via frictionless
 petl==1.7.15
+    # via frictionless
+pyarrow==17.0.0
     # via frictionless
 pydantic==2.8.2
     # via frictionless
@@ -58,9 +64,13 @@ pydantic-core==2.20.1
 pygments==2.18.0
     # via rich
 python-dateutil==2.9.0.post0
-    # via frictionless
+    # via
+    #   frictionless
+    #   pandas
 python-slugify==8.0.4
     # via frictionless
+pytz==2024.1
+    # via pandas
 pyyaml==6.0.2
     # via
     #   frictionless
@@ -103,6 +113,8 @@ typing-extensions==4.12.2
     #   pydantic
     #   pydantic-core
     #   typer
+tzdata==2024.1
+    # via pandas
 urllib3==2.2.2
     # via requests
 validators==0.33.0
